+++ VehicleInsurance/Interfaces/ICalculate.cs
@@ -4,6 +4,6 @@ namespace VehicleInsurance.Interfaces
 {
    public interface ICalculate
    {
-      double ImplementRule(Policy policy, double premium);
+      decimal ImplementRule(Policy policy, decimal premium);
    }
 }
